fix(automations): keep web chat URL helpers client-safe
Move path and URL builders out of the database module so the copy field can render without pulling server env into client tests.
